+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Changed
+
+-   The documentation has been revamped and expanded.
+
+### Removed
+
+-   OverlayOpened & OverlayClosed signals.
+
 ## [1.14.0-beta 2023-12-08]
 
 ### Added